1. Optimize plugin parameter stability during concurrent compilation 2. Add ASM write to FTAutoTrack.trackRequestBuilder method, requires ft-sdk version 1.6.10 or above 3. Support automatic capture of Log.println logs, requires ft-sdk version 1.6.8 or above 4. Fix issue where sourcemap symbol file was not generated when minifyEnabled was not enabled 5. Support generating sourcemap only without uploading by setting generateSourceMapOnly to true 6. Optimize automatic acquisition and upload of native symbol so files, support custom nativeLibPath 7. Translate to English

+++ ft-plugin/src/main/java/com/ft/plugin/garble/FTMapUploader.java
@@ -1,13 +1,11 @@
 package com.ft.plugin.garble;
 
 import com.android.build.gradle.AppExtension;
-import com.android.utils.FileUtils;
 
+import org.apache.tools.ant.util.FileUtils;
 import org.gradle.api.Project;
 import org.gradle.api.Task;
 import org.gradle.api.UnknownDomainObjectException;
-import org.gradle.api.artifacts.Configuration;
-import org.gradle.api.artifacts.ProjectDependency;
 
 import java.io.BufferedReader;
 import java.io.File;
@@ -18,49 +16,42 @@ import java.util.ArrayList;
 import java.util.HashMap;
 
 /**
- * 负责符号文件打包与上传
+ * Responsible for symbol file packaging and upload
  */
 public class FTMapUploader {
 
     private final Project project;
     /**
-     * debug symbol 路径
+     * so build symbol generation path
      */
-    private final static String CMAKE_DEBUG_SYMBOL_PATH = "/intermediates/cmake/debug/obj";
+    private final static String MERGED_LIB_PATH = "build/intermediates/merged_native_libs";
 
     /**
-     * native c/c++ build 生成路径
-     */
-    private final static String CMAKE_CXX_PATH = "/intermediates/cxx/Debug";
-
-    /**
-     * Unity Symbol 生成路径
+     * Unity Symbol generation path
      */
     private final static String UNITY_SYMBOLS_PATH = "/unityLibrary/symbols";
-    private final static String NAME_RELEASE_COMPILE_CLASSPATH = "releaseCompileClasspath";
 
     /**
-     * R8 混淆生成路径
+     * R8 obfuscation generation path
      */
     private final HashMap<String, ObfuscationSettingConfig> obfuscationSettingMap = new HashMap<>();
 
     /**
-     * 数据合并路径
+     * Data merge path
      */
     private final static String SYMBOL_MERGE_PATH_FORMAT = "/tmp/ftSourceMapMerge-%s";
     /**
-     * zip 文件打包路径
+     * zip file packaging path
      */
     private final static String SYMBOL_MERGE_ZIP_PATH_FORMAT = "/tmp/ftSourceMap-%s.zip";
     /**
-     * proguard 符号文件
+     * proguard symbol file
      */
     private final static String PROGUARD_MAPPING_PATH = "/outputs/proguard/%s/mapping/mapping.txt";
 
-    private final HashMap<String, ArrayList<String>> symbolPathsMap = new HashMap<>();
 
     /**
-     * zip 打包对象，临时生成路径
+     * zip packaging object, temporary generation path
      */
     private final String tmpBuildPathFormat;
     private final String zipBuildPathFormat;
@@ -80,11 +71,11 @@ public class FTMapUploader {
     }
 
     /**
-     * 上传混淆符号文件
+     * Upload obfuscation symbol file and native debug symbol file
      */
     public void configMapUpload() {
         if (flavorModelHashMap.isEmpty()) {
-            if (!extension.autoUploadMap) {
+            if (!extension.autoUploadMap && !extension.autoUploadNativeDebugSymbol) {
                 return;
             }
         }
@@ -96,45 +87,70 @@ public class FTMapUploader {
             String capVariantName = FTStringUtils.captitalizedString(variantName);
 
             String assembleTaskName = "assemble" + capVariantName;
+            Task zipTask = project.getTasks().create("ft" + capVariantName + "ZipSourceMap", task -> {
+                    })
+                    .doLast(task -> {
+                        ProductFlavorModel model = getFlavorModelFromName(variantName);
+                        Logger.debug("ZipSourceMap:" + model);
 
-            Task ftTask = project.getTasks().create("ft" + capVariantName + "UploadSymbolMap", task -> {
+                        if (!model.isAutoUploadMap() && !model.isAutoUploadNativeDebugSymbol()) {
+                            return;
+                        }
 
+                        ArrayList<String> symbolPaths = new ArrayList<>();
+                        if (model.isAutoUploadNativeDebugSymbol()) {
+                            appendSymbolPath(project, symbolPaths, variantName, model.getNativeLibPath());
+                        }
+
+                        String tmpBuildPath = String.format(tmpBuildPathFormat, variantName);
+                        String zipBuildPath = String.format(zipBuildPathFormat, variantName);
+                        model.setZipPath(zipBuildPath);
+
+                        // Delete previous cache
+                        try {
+                            deleteRecursively(new File(tmpBuildPath));
+                            FileUtils.delete(new File(zipBuildPath));
+
+                            ObfuscationSettingConfig config = obfuscationSettingMap.get(assembleTaskName);
+                            Logger.debug("task:" + assembleTaskName + ",config:" + config);
+                            if (config != null) {
+                                if (model.isAutoUploadNativeDebugSymbol()) {
+                                    if (!symbolPaths.isEmpty()) {
+                                        FTFileUtils.copyDifferentFolderFilesIntoOne(tmpBuildPath, symbolPaths.toArray(new String[0]));
+                                    } else {
+                                        Logger.error("not find native symbol path");
+                                    }
+                                }
+                                if (model.isAutoUploadMap()) {
+                                    if (!config.mappingOutputPath.isEmpty() && new File(config.mappingOutputPath).exists()) {
+                                        FTFileUtils.copyFile(new File(config.mappingOutputPath), new File(tmpBuildPath + "/mapping.txt"));
+                                    } else {
+                                        Logger.error("Mapping path empty or File not found");
+                                    }
+                                }
+                                FTFileUtils.zipFiles(new File(tmpBuildPath).listFiles(), new File(zipBuildPath));
+                            }
+                            Logger.debug("ZipSourceMap:" + assembleTaskName + " finish, zipPath:" + model.getZipPath());
+
+                        } catch (IOException e) {
+                            e.printStackTrace();
+                        }
+                    });
+
+            Task uploadTask = project.getTasks().create("ft" + capVariantName + "UploadSymbolMap", task -> {
             }).doLast(task -> {
-                ProductFlavorModel model = getFlavorModelFromName(variantName);
-                Logger.debug("ProductFlavorModel:" + model);
 
+                ProductFlavorModel model = getFlavorModelFromName(variantName);
                 if (!model.isAutoUploadMap() && !model.isAutoUploadNativeDebugSymbol()) {
                     return;
                 }
-
-                String tmpBuildPath = String.format(tmpBuildPathFormat, variantName);
-                String zipBuildPath = String.format(zipBuildPathFormat, variantName);
-                try {
-                    //删除之前的 cache
-                    FileUtils.deleteIfExists(new File(tmpBuildPath));
-                    FileUtils.deleteIfExists(new File(zipBuildPath));
-                } catch (IOException e) {
-                    Logger.debug("tmp  delete error:" + e.getMessage());
-                }
+                String zipBuildPath = model.getZipPath();
+                // Delete previous cache
                 try {
                     ObfuscationSettingConfig config = obfuscationSettingMap.get(assembleTaskName);
-                    Logger.debug("task:" + assembleTaskName + ",config:" + config + "");
                     if (config != null) {
-                        if (model.isAutoUploadNativeDebugSymbol()) {
-                            ArrayList<String> symbolPaths = symbolPathsMap.get(model.getName());
-                            if (symbolPaths != null && !symbolPaths.isEmpty()) {
-                                FTFileUtils.copyDifferentFolderFilesIntoOne(tmpBuildPath, symbolPaths.toArray(new String[0]));
-                            }
-                        }
-                        if (model.isAutoUploadMap()) {
-                            if (new File(config.mappingOutputPath).exists()) {
-                                FTFileUtils.copyFile(new File(config.mappingOutputPath), new File(tmpBuildPath + "/mapping.txt"));
-                            }
-                        }
-                        FTFileUtils.zipFiles(new File(tmpBuildPath).listFiles(), new File(zipBuildPath));
                         uploadWithParams(config, model, zipBuildPath);
                     }
-
                 } catch (IOException | InterruptedException e) {
                     e.printStackTrace();
                 }
@@ -143,7 +159,11 @@ public class FTMapUploader {
 
             if (!applicationVariant.getName().endsWith("Debug")) {
                 project.afterEvaluate(p -> {
-                    p.getTasks().getAt(assembleTaskName).finalizedBy(ftTask);
+                    ProductFlavorModel model = getFlavorModelFromName(variantName);
+                    p.getTasks().getAt(assembleTaskName).finalizedBy(zipTask);
+                    if (!model.isGenerateSourceMapOnly()) {
+                        p.getTasks().getAt(assembleTaskName).finalizedBy(uploadTask);
+                    }
                 });
             }
 
@@ -184,6 +204,9 @@ public class FTMapUploader {
                                     variantName);
                             Logger.debug("Map Config:" + config + ",task:" + task.getName());
                             obfuscationSettingMap.put(task.getName(), config);
+                        } else {
+                            obfuscationSettingMap.put(task.getName(), new ObfuscationSettingConfig());
+                            Logger.error("MinifyEnabled or Proguard Setting not found");
                         }
                     }
                 }
@@ -192,98 +215,69 @@ public class FTMapUploader {
 
     }
 
-
     /**
-     * 上传 debug native symbol 文件
+     * Delete folder
+     *
+     * @param file
+     * @throws IOException
      */
-    public void configNativeSymbolUpload() {
-        if (flavorModelHashMap.isEmpty()) {
-            if (!extension.autoUploadNativeDebugSymbol) {
-                return;
+    private void deleteRecursively(File file) {
+        if (file.isDirectory()) {
+            File[] fileList = file.listFiles();
+            if (fileList != null) {
+                for (File subFile : fileList) {
+                    deleteRecursively(subFile);
+                }
             }
         }
-        project.afterEvaluate(p -> {
-            AppExtension appExtension = (AppExtension) p.getProperties().get("android");
-
-            if (appExtension.getProductFlavors().size() > 0) {
-                appExtension.getProductFlavors().all(flavor -> {
-                    // 获取 Flavor 的名称
-                    String flavorName = flavor.getName();
-                    ArrayList<String> symbolPaths = symbolPathsMap.computeIfAbsent(flavorName, k -> new ArrayList<>());
-                    appendSymbolPath(p, symbolPaths, flavorName);
-
-                });
-            } else {
-                ArrayList<String> symbolPaths = symbolPathsMap.computeIfAbsent("release", k -> new ArrayList<>());
-                appendSymbolPath(p, symbolPaths, "");
-            }
-
-
-        });
-
+        FileUtils.delete(file);  // Use FileUtils to delete file or folder
     }
 
+
     /**
-     * 扫描并添加 native cmake build 路径
+     * Scan and add native cmake build path
      *
      * @param p
      * @param list
-     * @param flavor
+     * @param nativeLibPath
      */
-    private void appendSymbolPath(Project p, ArrayList<String> list, String flavor) {
+    private void appendSymbolPath(Project p, ArrayList<String> list, String variantName, String nativeLibPath) {
 
         p.getAllprojects().forEach(subProject -> {
-            String buildPath = subProject.getBuildDir().getAbsolutePath();
-            String debugSymbolPath = buildPath + CMAKE_DEBUG_SYMBOL_PATH;
-
+            File projectPath = subProject.getProjectDir();
+            String debugSymbolPath = "";
+            if (nativeLibPath.isEmpty()) {
+                debugSymbolPath = new File(projectPath, MERGED_LIB_PATH + "/" + variantName + "/out/lib").getAbsolutePath();
+            } else {
+                debugSymbolPath = new File(projectPath, nativeLibPath).getAbsolutePath();
+            }
             File file = new File(debugSymbolPath);
             if (file.exists()) {
+                Logger.debug("debugSymbolPath:" + debugSymbolPath);
                 list.add(debugSymbolPath);
             }
         });
-        String name = flavor.length() == 0 ? NAME_RELEASE_COMPILE_CLASSPATH :
-                (flavor + FTStringUtils.captitalizedString(NAME_RELEASE_COMPILE_CLASSPATH));
-        Configuration configuration = p.getConfigurations().findByName(name);
-        if (configuration != null) {
-            String rootPath = p.getRootDir().getAbsolutePath();
-            configuration.getAllDependencies().forEach(dependency -> {
-                if (dependency instanceof ProjectDependency) {
-                    String moduleName = dependency.getName();
-                    String buildPath = rootPath + "/" + moduleName + "/build";
-                    String debugSymbolPath = buildPath + CMAKE_DEBUG_SYMBOL_PATH;
-                    File file = new File(debugSymbolPath);
-                    if (file.exists()) {
-                        Logger.debug("debugSymbolPath:" + debugSymbolPath);
-                        list.add(debugSymbolPath);
-                    }
 
-                    String unitySymbolsPath = rootPath + UNITY_SYMBOLS_PATH;
-                    File unitySymbols = new File(unitySymbolsPath);
-                    Logger.debug("unitySymbolsPath:" + unitySymbolsPath);
-                    if (unitySymbols.exists()) {
-                        list.add(unitySymbolsPath);
-                    }
-                }
-            });
-            if (symbolPathsMap.isEmpty()) {
-                Logger.error("native symbol not found");
-            } else {
-                Logger.debug("paths:" + symbolPathsMap);
-            }
-
+        //unity symbol
+        String rootPath = p.getRootDir().getAbsolutePath();
+        String unitySymbolsPath = rootPath + UNITY_SYMBOLS_PATH;
+        File unitySymbols = new File(unitySymbolsPath);
+        if (unitySymbols.exists()) {
+            Logger.debug("unitySymbolsPath:" + unitySymbolsPath);
+            list.add(unitySymbolsPath);
         }
-
     }
 
+
     /**
-     * 上传符号文件
+     * Upload symbol file
      *
-     * @param settingConfig 混淆配置
+     * @param settingConfig  obfuscation configuration
      * @throws IOException
      * @throws InterruptedException
      */
     private void uploadWithParams(ObfuscationSettingConfig settingConfig, ProductFlavorModel model, String zipBuildPath) throws IOException, InterruptedException {
-        Logger.debug(model.toString());
+        Logger.debug("uploadWithParams:" + model.toString());
         String cmd = "curl -m 1800 -X PUT " + model.getDatakitUrl() + "/v1/sourcemap?"
                 + "app_id=" + model.getAppId()
                 + "&env=" + model.getEnv()
@@ -312,7 +306,7 @@ public class FTMapUploader {
 
         int exitCode = process.exitValue();
         if (exitCode != 0) {
-            Logger.error("map 文件上传失败");
+            Logger.error("map file upload failed");
             Logger.error("exit code::" + exitCode);
         }
 
@@ -323,11 +317,11 @@ public class FTMapUploader {
         ProductFlavorModel model = flavorModelHashMap
                 .get(variantName.replace("Release", ""));
         if (model != null) {
-            //多个 flavor
+            //multiple flavor
             model.mergeFTExtension(extension);
             return model;
         } else {
-            //不设置 flavor, 此处 variantName = release
+            //not set flavor, here variantName = release
             model = new ProductFlavorModel(variantName);
             model.setFromFTExtension(extension);
         }
@@ -336,25 +330,25 @@ public class FTMapUploader {
 
 
     /**
-     * 混淆配置
+     * obfuscation configuration
      */
     static class ObfuscationSettingConfig {
         /**
-         * 包名
+         * Package name
          */
-        String applicationId;
+        String applicationId = "";
         /**
-         * 版本号 字符 例如 1.0.0
+         * Version name, e.g. 1.0.0
          */
-        String versionName;
+        String versionName = "";
         /**
          * Build Code
          */
         int versionCode;
         /**
-         * map 输出地址
+         * map output path
          */
-        String mappingOutputPath;
+        String mappingOutputPath = "";
 
         @Override
         public String toString() {
